refactor(games): extract cycle helpers and sponsors from games/mod.rs
